Allow aggressive consumers to join the convergence mechanism.
If aggressive consumers (those that raise exceptions if their dependencies are
not satisfied) fail they will now be picked up by the convergence mechanism
and re-tried.

When a working set does not converge we will also show all errors that
occurred in the working set before exiting (by logging on error and exception
level).