Rework dealing with circular resource dependencies:
From now on, we prohibit circular resource dependencies.

Most cases circular references can be dealt with by splitting up components,
and using the 'settings' or 'secrets' pattern as a way to extract pure data
dependencies.

We thus removed the data-based approach of trying to find a stable equilibrium
as this proved sufficienlty complicated (either troubles with references or
restricting the key/value data model to simple data structures) and limiting
the expressiveness of component code.

We now fail explicitly when detecting cycles and try to give a readable message
and snapshot of the dependency graph that shows where the cycle was found.