Refactor passphrase acquisition to allow keeping a passphrase cache.
The passphrase cache is only intended for batou-local runs in AutoMode and
should not be intermingled with batou-remote passphrases.

I also removed a superfluous abstraction and made the secrets component a
little bit more special so it appears in some of the core code now. Hopefully
this will allow a better abstraction in the future. Also, the 'set' remote
command is gone now.

Fixes #11413