Finish massive restructuring.
Internal structures do not model the 'service' any longer, as it was really
painful. Also, the RootComponents have been cleaned up and this resulted in a
few nice side-effects (like working directory management) that prepare us for
the future.

Also a lot of bugfixes along the way as I cleaned up the tests.