fix: SyncDirectory Prediction
Add a check to the SyncDirectory component for the target directory.

The SyncDirectory component `verify()` stage fails when the target
directory does not exist yet, because it uses rsync to determine if there
is a diff between left and right. In cases where the target directory
is part of the deployment and has thusly not been created yet, this
leads to unexpected errors.