improve secret handling performance
Massively improve secret handling performance by avoiding large
amounts of writing to / reading from temporary files.

This solves a lot of complexity around the age call because it
does not require piping the ssh key's password into a subprocess' stdin
conditionally.

Instead, this can be handled using `cryptography` to handle the
password-protected ssh key and `pyrage`, python bindings to the rage
library which supports de/encrypting in-memory.

Another benefit of this implementation is a relaxed requirement (now
optional) on the `age` CLI since the bindings to the rage library
provide a full implementation out of the box.